requirements.txt
Flask-JWT-Extended
Flask-RESTful
flask_cors
requests
python-decouple
Flask-SQLAlchemy
schedule
Flask-APScheduler